quadnet_mimu/requirements.txt
torch>=1.12
numpy>=1.22
pandas>=1.4
matplotlib>=3.6
tensorboard>=2.9
tqdm>=4.64
scikit-learn>=1.1
pyyaml>=6.0
scipy>=1.8
einops>=0.8.0

